// src/features/editor/types.ts
export const fonts = [
  "Arial",
  "Arial Black",
  "Verdana",
  "Helvetica",
  "Tahoma",
  "Trebuchet MS",
  "Times New Roman",
  "Georgia",
  "Garamond",
  "Courier New",
  "Brush Script MT",
  "Palatino",
  "Bookman",
  "Comic Sans MS",
  "Impact",
  "Lucida Sans Unicode",
  "Geneva",
  "Lucida Console",
];

export const selectionDependentTools = [
  "fill",
  "font",
  "filter",
  "opacity",
  "remove-bg",
  "stroke-color",
  "stroke-width",
];
// ...
export type ActiveTool =
  | "select"
  | "shapes"
  | "text"
  | "images"
  | "draw"
  | "fill"
  | "stroke-color"
  | "stroke-width"
  | "font"
  | "opacity"
  | "filter"
  | "settings"
  | "ai"
  | "content"
  | "remove-bg"
  | "templates";
// ...
export type Font = typeof fonts[number];

export interface Narrative {
  productName: string;
  index: number;
  title: {
    text: string;
    vertical_align: 'top' | 'middle' | 'bottom';
    horizontal_align: 'left' | 'right' | 'center';
    font: Font;
  };
  tagline: {
    text: string;
    vertical_align: 'top' | 'middle' | 'bottom';
    horizontal_align: 'left' | 'right' | 'center';
    font: Font;
  };
  image: {
    background: string;
    description: string;
    product: {
      vertical_align: 'top' | 'middle' | 'bottom';
      horizontal_align: 'left' | 'right' | 'center';
    };
    logo: {
      vertical_align: 'top' | 'middle' | 'bottom';
      horizontal_align: 'left' | 'right' | 'center';
    };
  };
}